Rename `debanders` to `funcs`, update docs
