--- docs/submodules/debanders.rst
@@ -1,16 +1,0 @@
-=========
-Functions
-=========
-
-.. autosummary::
-
-    vsdeband.debanders.dumb3kdb
-    vsdeband.debanders.f3kbilateral
-    vsdeband.debanders.f3kpf
-    vsdeband.debanders.lfdeband
-    vsdeband.debanders.placebo_deband
-
-.. automodule:: vsdeband.debanders
-    :members:
-    :undoc-members:
-    :show-inheritance:
