+++ docs/submodules/f3kdb.rst
@@ -5,6 +5,7 @@ f3kdb
 .. autosummary::
 
     vsdeband.f3kdb.SampleMode
+    vsdeband.f3kdb.F3kdbPlugin
     vsdeband.f3kdb.F3kdb
 
 .. automodule:: vsdeband.f3kdb
