+++ docs/submodules/funcs.rst
@@ -1,0 +1,14 @@
+=========
+Functions
+=========
+
+.. autosummary::
+
+    vsdeband.funcs.f3kbilateral
+    vsdeband.funcs.f3kpf
+    vsdeband.funcs.lfdeband
+
+.. automodule:: vsdeband.funcs
+    :members:
+    :undoc-members:
+    :show-inheritance:
