+++ docs/submodules/placebo.rst
@@ -4,6 +4,7 @@ Placebo
 
 .. autosummary::
 
+    vsdeband.placebo.PlaceboDither
     vsdeband.placebo.Placebo
 
 .. automodule:: vsdeband.placebo
