+++ vsdeband/__init__.py
@@ -1,4 +1,4 @@
 from .abstract import *  # noqa: F401, F403
-from .debanders import *  # noqa: F401, F403
+from .funcs import *  # noqa: F401, F403
 from .f3kdb import *  # noqa: F401, F403
 from .placebo import *  # noqa: F401, F403
