+++ vsdeband/funcs.py
